Created base class for scrapers
Now all scrapers inherit from the base class and implement the
basic function for all following added scrapers.
Also did some cleaning up of the code and added some comments.

Scrapers need to be added to the __init__.py file in the scrapers
and implement the @register_scraper() decorator to be added to the
scraper registry.